Refactor StockEntity
