+++ core-stock/src/main/java/com/example/demo/core/mapper/StockMapper.java
@@ -1,6 +1,6 @@
 package com.example.demo.core.mapper;
 
-import com.example.demo.core.repository.model.StockModel;
+import com.example.demo.core.repository.entity.StockEntity;
 import com.example.demo.dto.common.StockShoe;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
@@ -14,15 +14,15 @@ public interface StockMapper {
     @Mapping(source = "stockId.size", target = "size")
     @Mapping(source = "stockId.color", target = "color")
     @Mapping(source = "quantity", target = "quantity")
-    StockShoe toStockShoe (StockModel stock);
+    StockShoe toStockShoe (StockEntity stock);
 
     @Mapping(source = "name", target = "stockId.name")
     @Mapping(source = "size", target = "stockId.size")
     @Mapping(source = "color", target = "stockId.color")
     @Mapping(source = "quantity", target = "quantity")
-    StockModel toStock (StockShoe stockShoe);
+    StockEntity toStock (StockShoe stockShoe);
 
-    List<StockShoe> toListStockShoes(List<StockModel> stock);
+    List<StockShoe> toListStockShoes(List<StockEntity> stock);
 
-    List<StockModel> toListStock(List<StockShoe> stockShoes);
+    List<StockEntity> toListStock(List<StockShoe> stockShoes);
 }
