+++ core-stock/src/main/java/com/example/demo/core/repository/StockRepository.java
@@ -1,11 +1,11 @@
 package com.example.demo.core.repository;
 
-import com.example.demo.core.repository.model.StockModel;
-import com.example.demo.core.repository.model.StockId;
+import com.example.demo.core.repository.entity.StockEntity;
+import com.example.demo.core.repository.entity.StockId;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StockRepository extends MongoRepository<StockModel, StockId> {
+public interface StockRepository extends MongoRepository<StockEntity, StockId> {
 
 }
